Use strict language for node and component
In solph, the term "component" is strictly used for a type
of node that has customisable relations between input and
output flows (in contrast to buses that just maintain
energy balance). To not confuse people, we should adhere to
that nomencature.

The "AbstractSolphComponent" always was a collection of solph
nodes, so it is renamed to "AbstractSolphRepresentation".